tests/wpt/web-platform-tests/css/css-color/parsing/opacity-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Grid Layout Test: getComputedStyle().opacity</title>
<link rel="help" href="https://www.w3.org/TR/css-color-3/#opacity">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("opacity", "1");
test_computed_value("opacity", "0.5");
test_computed_value("opacity", "0");
test_computed_value("opacity", "-2", "0");
test_computed_value("opacity", "3", "1");
test_computed_value("opacity", "-100%", "0");
test_computed_value("opacity", "50%", "0.5");
test_computed_value("opacity", "300%", "1");

</script>
</body>
</html>
